Add dragg navigation
