+++ package.json
@@ -27,6 +27,7 @@
     "react": "17.0.1",
     "react-audio-player-pro": "^1.1.9",
     "react-dom": "17.0.1",
+    "react-indiana-drag-scroll": "^1.8.1",
     "react-is": "^17.0.2",
     "react-pdf": "^5.2.0",
     "styled-components": "^5.2.1",
